refactor: extract shared TaskCard and page-specific hooks
- Create shared TaskCard component supporting backlog/sprint/kanban/compact variants
- Extract sprint page logic into useSprintTasks and useSprintManagement hooks
- Extract settings page logic into useAccount, useThresholds, useMemory, useMemoryQuestions hooks
- Reduce god component pattern across page files
